Match more types and format specifiers to size_t fontdata_used
Omissions in commit 910211f1c0078e37546f73e95306724358b89be2.

Microsoft's C library does not support %zu until Universal CRT
(Visual Studio 2015). At worst, this verbose-level message will
look wrong and be useless.

+++ libass/ass.c
@@ -820,7 +820,7 @@ static unsigned char *decode_chars(const unsigned char *src,
                                    unsigned char *dst, size_t cnt_in)
 {
     uint32_t value = 0;
-    for (int i = 0; i < cnt_in; i++)
+    for (size_t i = 0; i < cnt_in; i++)
         value |= (uint32_t) ((src[i] - 33u) & 63) << 6 * (3 - i);
 
     *dst++ = value >> 16;
@@ -850,7 +850,7 @@ static int decode_font(ASS_Track *track)
     size_t dsize;                  // decoded size
     unsigned char *buf = 0;
 
-    ass_msg(track->library, MSGL_V, "Font: %d bytes encoded data",
+    ass_msg(track->library, MSGL_V, "Font: %zu bytes encoded data",
             track->parser_priv->fontdata_used);
     size = track->parser_priv->fontdata_used;
     if (size % 4 == 1) {
